Make it easier to create mappers on `PageTable`
Store the physical mapping base on the `PageTable` instead of just on the
`Mapper`. This means we don't have to provide it every time we want to
touch a page table. I thought this would be more confusing, but it's fine
and much more convenient.

+++ lib/x86_64/src/memory/kernel_map.rs
@@ -21,7 +21,7 @@ pub fn physical_to_virtual(address: PhysicalAddress) -> VirtualAddress {
 }
 
 /// The kernel is mapped into the 511th entry of the P4.
-pub const KERNEL_P4_ENTRY: u16 = 511;
+pub const KERNEL_P4_ENTRY: usize = 511;
 
 pub const KERNEL_ADDRESS_SPACE_START: VirtualAddress =
     unsafe { VirtualAddress::new_unchecked(0xffff_ff80_0000_0000) };
